+++ backend/utils/ini-reader.go
@@ -25,6 +25,7 @@ type EMLyConfig struct {
 	UpdateAutoCheck          string `ini:"UPDATE_AUTO_CHECK"`
 	BugReportAPIURL          string `ini:"BUGREPORT_API_URL"`
 	BugReportAPIKey          string `ini:"BUGREPORT_API_KEY"`
+	LogLevel                 string `ini:"LOG_LEVEL"`
 }
 
 // LoadConfig reads the config.ini file at the given path and returns a Config struct