+++ main.go
@@ -6,6 +6,8 @@ import (
 	"os"
 	"strings"
 
+	pkglogger "emly/backend/logger"
+
 	"github.com/wailsapp/wails/v2"
 	"github.com/wailsapp/wails/v2/pkg/options"
 	"github.com/wailsapp/wails/v2/pkg/options/assetserver"
@@ -19,11 +21,12 @@ func (a *App) onSecondInstanceLaunch(secondInstanceData options.SecondInstanceDa
 	var secondInstanceArgs []string
 	secondInstanceArgs = secondInstanceData.Args
 
-	Log("user opened second instance", strings.Join(secondInstanceData.Args, ","))
-	Log("user opened second from", secondInstanceData.WorkingDirectory)
+	pkglogger.Info("second instance launched",
+		"args", strings.Join(secondInstanceData.Args, ","),
+		"working_dir", secondInstanceData.WorkingDirectory,
+	)
 	runtime.WindowUnminimise(a.ctx)
 	runtime.WindowShow(a.ctx)
-	Log("launchArgs", secondInstanceArgs)
 	go runtime.EventsEmit(a.ctx, "launchArgs", secondInstanceArgs)
 }
 
@@ -43,9 +46,6 @@ func main() {
 
 	for _, arg := range args {
 		if strings.Contains(arg, "--view-image") {
-			uniqueId = "emly-viewer-" + strings.ReplaceAll(arg, "--view-image=", "") // Make unique per image or just random?
-			// Actually, just using a different base ID allows multiple viewers if we append something random or just use "mailpaw-viewer" and disable single instance for viewers?
-			// Let's just disable single instance for viewers by generating a random ID or appending timestamp
 			uniqueId = "emly-viewer-" + arg // simplified uniqueness
 			windowTitle = "EMLy Image Viewer"
 			windowWidth = 800
@@ -97,6 +97,6 @@ func main() {
 	})
 
 	if err != nil {
-		println("Error:", err.Error())
+		pkglogger.Error("application error", "error", err.Error())
 	}
 }